backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
passlib[bcrypt]
pyjwt
beautifulsoup4
requests
lxml
trafilatura
newspaper3k
vaderSentiment
textstat
web3
PyPDF2
python-docx
